Fixed info to terminal,
Gjort sådan at de to floatstr bliver lagret og printet som 2 seperate strings, så acc ikke bliver overskrevet og printet som hast.

+++ Hurtig-Bil/Hurtig-Bil/Speedometer.h
@@ -24,8 +24,8 @@ struct Bil_t
     float OldHastighed;     // Den forrige_hastighedsmåling
 	float Hastighed;        // Den aktuelle_hastighedsmåling
 	float Acceleration;     // Hastighedsændring_= Acc/Deceleration
-	int   Hast;             // Hastighed som_integer
-	int   Acc;              // Accelleration_som_integer
+	float   Hast;             // Hastighed som_integer
+	float   Acc;              // Accelleration_som_integer
 	int   Odo;              // Afstandsmåler_til Speedometer
 } Bil;
